Typescriptify Applications-Component
